Streamline CLI and document usage
