--- __main__.py
@@ -1,5 +1,0 @@
-# grimbrain/__main__.py
-from .cli import main
-
-if __name__ == "__main__":
-    raise SystemExit(main())
